refactor node-demos to utilize a filesystem module
I've currently checked in the node_modules folder in `demos/nodejs` as
it is only a symbolic link to the root of the repo. I'm unsure if this
is supported cross platform, specifically windows.

The advantage here is that the demo scripts now require("flocking"),
which is much nicer to copy paste 🎉